refactor llm, async completion, streaming, retries
